adds sign and upload for 3rd party artifacts
`sign-3p-artifacts` is now fully implemented. It may need a fair amount of testing -- the code is ugly. We may want to implement this in a POM instead of via Bash at some point soon in the future, since Maven is likely to be better working with Maven artifacts than Bash having to reverse-engineer the naming convention.