Add handling of signed POM + unsigned artifacts
for some reason, `xercesImpl` version `2.9.1` has a signed POM, but no signature on its JAR. this seems to completely defeat the purpose of having a signed artifact at all, since the JAR is the important part. either way, we need to be able to sign all these artifacts ourselves to make pgpVerify happy in our own builds, so now we have handling for this extremely special case.