Eliminate the `--provider` CLI Option
Each command is written to deal with a specific provider. Most of our artifacts go to JFrog, with the exception of binary artifacts FR built that we're archiving to BinTray. As such, it doesn't really make sense to allow a `--provider` CLI option when most of our commands don't support it.