Modify Deploy Tool to Work with `wrensec-pgp-whitelist`
`wrensec-home` no longer maintains the PGP whitelist; it's now a separate Maven artifact. This change modifies the Wren Deploy tool to work with that new project structure instead.

In the new structure, `wrensec-parent` forces the whitelist location to
be provided by the new `pgpVerifyWhitelist` property, which we have to
set since `pgpverify.keysMapLocation` gets overridden by the pom.xml in
the current working directory.