Automatically Detect when to Amend the Whitelist
This eliminates the need to remember to use the `--amend` option of `capture-unapproved-artifact-sigs`. The original option has been renamed `--force-amend` for cases in which the latest commit of the PGP whitelist project does not match the auto-generated subject line for the current project and project version, but amending is still desired.