Automatically Sort and De-dupe the PGP Whitelist
Previously, the `capture-unapproved-artifact-sigs` sub-command would append all new deps to the end of the file. Later, before the new whitelist was released, someone (usually me) would have to manually re-sort the whitelist so that the deps are grouped together. Now, this manual effort is no longer required!

Via some magic with `sed`, `sort`, and `uniq` the sub-command is now intelligent enough to sort and de-dupe the whitelist in-place before committing.